[PageBlob]Optimize Download Sparse File
