+++ sdk/storage/azure-storage-blob/tests/test_page_blob_async.py
@@ -119,6 +119,18 @@ class StoragePageBlobTestAsync(StorageTestCase):
         await blob_client.upload_page(data, start_range, end_range)
         return blob_client
 
+    async def _create_sparse_page_blob(self, size=1024*1024, data=''):
+        blob_client = self._get_blob_reference()
+        await blob_client.create_page_blob(size=size)
+
+        range_start = 8*1024 + 512
+        range_end = range_start + len(data) - 1
+
+        # the page blob will be super sparse like this:'                         some data                      '
+        await blob_client.upload_page(data, range_start, range_end)
+
+        return blob_client
+
     async def _wait_for_async_copy(self, blob):
         count = 0
         props = await blob.get_blob_properties()
@@ -1833,6 +1845,38 @@ class StoragePageBlobTestAsync(StorageTestCase):
         loop = asyncio.get_event_loop()
         loop.run_until_complete(self._test_blob_tier_copy_blob())
 
+    async def _test_download_sparse_page_blob(self):
+        # Arrange
+        await self._setup()
+        self.config.max_single_get_size = 4*1024
+        self.config.max_chunk_get_size = 1024
+
+        sparse_page_blob_size = 1024 * 1024
+        data = self.get_random_bytes(2048)
+        blob_client = await self._create_sparse_page_blob(size=sparse_page_blob_size, data=data)
+
+        # Act
+        page_ranges, cleared = await blob_client.get_page_ranges()
+        start = page_ranges[0]['start']
+        end = page_ranges[0]['end']
+
+        content = await blob_client.download_blob()
+        content = await content.content_as_bytes()
+
+        # Assert
+        self.assertEqual(sparse_page_blob_size, len(content))
+        # make sure downloaded data is the same as the uploaded data
+        self.assertEqual(data, content[start: end + 1])
+        # assert all unlisted ranges are empty
+        for byte in content[:start-1]:
+            self.assertEqual(byte, 0)
+        for byte in content[end+1:]:
+            self.assertEqual(byte, 0)
+
+    @record
+    def test_download_sparse_page_blob_async(self):
+        loop = asyncio.get_event_loop()
+        loop.run_until_complete(self._test_download_sparse_page_blob())
 
 # ------------------------------------------------------------------------------
 if __name__ == '__main__':
