+++ sdk/storage/azure-storage-file/azure/storage/file/aio/file_client_async.py
@@ -22,7 +22,6 @@ from .._generated.version import VERSION
 from .._generated.models import StorageErrorException, FileHTTPHeaders
 from .._shared.policies_async import ExponentialRetry
 from .._shared.uploads_async import upload_data_chunks, FileChunkUploader, IterStreamer
-from .._shared.downloads_async import StorageStreamDownloader
 from .._shared.base_client_async import AsyncStorageAccountHostsMixin
 from .._shared.request_handlers import add_metadata_headers, get_length
 from .._shared.response_handlers import return_response_headers, process_storage_error
@@ -30,6 +29,7 @@ from .._deserialize import deserialize_file_properties, deserialize_file_stream
 from ..file_client import FileClient as FileClientBase
 from ._polling_async import CloseHandlesAsync
 from .models import HandlesPaged
+from .download_async import StorageStreamDownloader
 
 if TYPE_CHECKING:
     from datetime import datetime
@@ -463,7 +463,7 @@ class FileClient(AsyncStorageAccountHostsMixin, FileClientBase):
             raise ValueError("Offset value must not be None if length is set.")
 
         downloader = StorageStreamDownloader(
-            service=self._client.file,
+            client=self._client.file,
             config=self._config,
             offset=offset,
             length=length,
