+++ sdk/storage/azure-storage-file/azure/storage/file/download.py
@@ -5,56 +5,75 @@
 # --------------------------------------------------------------------------
 
 import sys
-import asyncio
+import threading
 from io import BytesIO
-from itertools import islice
 
-from azure.core.exceptions import HttpResponseError
-
-from .request_handlers import validate_and_format_range_headers
-from .response_handlers import process_storage_error, parse_length_from_content_range
-from .encryption import decrypt_blob
-from .downloads import process_range_and_offset
+from azure.core import HttpResponseError
+from azure.core.tracing.common import with_current_context
+from ._shared.encryption import decrypt_blob
+from ._shared.request_handlers import validate_and_format_range_headers
+from ._shared.response_handlers import process_storage_error, parse_length_from_content_range
 
 
-async def process_content(data, start_offset, end_offset, encryption):
+def process_range_and_offset(start_range, end_range, length, encryption):
+    start_offset, end_offset = 0, 0
+    if encryption.get("key") is not None or encryption.get("resolver") is not None:
+        if start_range is not None:
+            # Align the start of the range along a 16 byte block
+            start_offset = start_range % 16
+            start_range -= start_offset
+
+            # Include an extra 16 bytes for the IV if necessary
+            # Because of the previous offsetting, start_range will always
+            # be a multiple of 16.
+            if start_range > 0:
+                start_offset += 16
+                start_range -= 16
+
+        if length is not None:
+            # Align the end of the range along a 16 byte block
+            end_offset = 15 - (end_range % 16)
+            end_range += end_offset
+
+    return (start_range, end_range), (start_offset, end_offset)
+
+
+def process_content(data, start_offset, end_offset, encryption):
     if data is None:
         raise ValueError("Response cannot be None.")
-    content = data.response.body()
-    if encryption.get('key') is not None or encryption.get('resolver') is not None:
+    content = b"".join(list(data))
+    if content and encryption.get("key") is not None or encryption.get("resolver") is not None:
         try:
             return decrypt_blob(
-                encryption.get('required'),
-                encryption.get('key'),
-                encryption.get('resolver'),
+                encryption.get("required"),
+                encryption.get("key"),
+                encryption.get("resolver"),
                 content,
                 start_offset,
                 end_offset,
-                data.response.headers)
+                data.response.headers,
+            )
         except Exception as error:
-            raise HttpResponseError(
-                message="Decryption failed.",
-                response=data.response,
-                error=error)
+            raise HttpResponseError(message="Decryption failed.", response=data.response, error=error)
     return content
 
 
-class _AsyncChunkDownloader(object):  # pylint: disable=too-many-instance-attributes
-
+class _ChunkDownloader(object):
     def __init__(
-            self, service=None,
-            total_size=None,
-            chunk_size=None,
-            current_progress=None,
-            start_range=None,
-            end_range=None,
-            stream=None,
-            parallel=None,
-            validate_content=None,
-            encryption_options=None,
-            **kwargs):
+        self,
+        client=None,
+        total_size=None,
+        chunk_size=None,
+        current_progress=None,
+        start_range=None,
+        end_range=None,
+        stream=None,
+        validate_content=None,
+        encryption_options=None,
+        **kwargs
+    ):
 
-        self.service = service
+        self.client = client
 
         # information on the download range/chunk size
         self.chunk_size = chunk_size
@@ -64,12 +83,6 @@ class _AsyncChunkDownloader(object):  # pylint: disable=too-many-instance-attrib
 
         # the destination that we will write to
         self.stream = stream
-        self.stream_lock = asyncio.Lock() if parallel else None
-        self.progress_lock = asyncio.Lock() if parallel else None
-
-        # for a parallel download, the stream is always seekable, so we note down the current position
-        # in order to seek to the right place when out-of-order chunks come in
-        self.stream_start = stream.tell() if parallel else None
 
         # download progress so far
         self.progress_total = current_progress
@@ -94,60 +107,109 @@ class _AsyncChunkDownloader(object):  # pylint: disable=too-many-instance-attrib
             yield index
             index += self.chunk_size
 
-    async def process_chunk(self, chunk_start):
+    def process_chunk(self, chunk_start):
         chunk_start, chunk_end = self._calculate_range(chunk_start)
-        chunk_data = await self._download_chunk(chunk_start, chunk_end)
+        chunk_data = self._download_chunk(chunk_start, chunk_end)
         length = chunk_end - chunk_start
         if length > 0:
-            await self._write_to_stream(chunk_data, chunk_start)
-            await self._update_progress(length)
+            self._write_to_stream(chunk_data, chunk_start)
+            self._update_progress(length)
 
-    async def yield_chunk(self, chunk_start):
+    def yield_chunk(self, chunk_start):
         chunk_start, chunk_end = self._calculate_range(chunk_start)
-        return await self._download_chunk(chunk_start, chunk_end)
+        return self._download_chunk(chunk_start, chunk_end)
 
-    async def _update_progress(self, length):
-        if self.progress_lock:
-            async with self.progress_lock:  # pylint: disable=not-async-context-manager
-                self.progress_total += length
-        else:
-            self.progress_total += length
+    # should be provided by the subclass
+    def _update_progress(self, length):
+        pass
 
-    async def _write_to_stream(self, chunk_data, chunk_start):
-        if self.stream_lock:
-            async with self.stream_lock:  # pylint: disable=not-async-context-manager
-                self.stream.seek(self.stream_start + (chunk_start - self.start_index))
-                self.stream.write(chunk_data)
-        else:
-            self.stream.write(chunk_data)
+    # should be provided by the subclass
+    def _write_to_stream(self, chunk_data, chunk_start):
+        pass
 
-    async def _download_chunk(self, chunk_start, chunk_end):
+    def _download_chunk(self, chunk_start, chunk_end):
         download_range, offset = process_range_and_offset(
-            chunk_start, chunk_end, chunk_end, self.encryption_options)
+            chunk_start, chunk_end, chunk_end, self.encryption_options
+        )
         range_header, range_validation = validate_and_format_range_headers(
-            download_range[0],
-            download_range[1] - 1,
-            check_content_md5=self.validate_content)
+            download_range[0], download_range[1] - 1, check_content_md5=self.validate_content
+        )
 
         try:
-            _, response = await self.service.download(
+            _, response = self.client.download(
                 range=range_header,
                 range_get_content_md5=range_validation,
                 validate_content=self.validate_content,
                 data_stream_total=self.total_size,
                 download_stream_current=self.progress_total,
-                **self.request_options)
+                **self.request_options
+            )
         except HttpResponseError as error:
             process_storage_error(error)
 
-        chunk_data = await process_content(response, offset[0], offset[1], self.encryption_options)
+        chunk_data = process_content(response, offset[0], offset[1], self.encryption_options)
 
         # This makes sure that if_match is set so that we can validate
         # that subsequent downloads are to an unmodified blob
-        if self.request_options.get('modified_access_conditions'):
-            self.request_options['modified_access_conditions'].if_match = response.properties.etag
-
+        if self.request_options.get("modified_access_conditions"):
+            self.request_options["modified_access_conditions"].if_match = response.properties.etag
         return chunk_data
+
+
+class ParallelChunkDownloader(_ChunkDownloader):
+    def __init__(
+        self,
+        client=None,
+        total_size=None,
+        chunk_size=None,
+        current_progress=None,
+        start_range=None,
+        end_range=None,
+        stream=None,
+        validate_content=None,
+        encryption_options=None,
+        **kwargs
+    ):
+        super(ParallelChunkDownloader, self).__init__(
+            client=client,
+            total_size=total_size,
+            chunk_size=chunk_size,
+            current_progress=current_progress,
+            start_range=start_range,
+            end_range=end_range,
+            stream=stream,
+            validate_content=validate_content,
+            encryption_options=encryption_options,
+            **kwargs
+        )
+
+        # for a parallel download, the stream is always seekable, so we note down the current position
+        # in order to seek to the right place when out-of-order chunks come in
+        self.stream_start = stream.tell()
+
+        # since parallel operations are going on
+        # it is essential to protect the writing and progress reporting operations
+        self.stream_lock = threading.Lock()
+        self.progress_lock = threading.Lock()
+
+    def _update_progress(self, length):
+        with self.progress_lock:
+            self.progress_total += length
+
+    def _write_to_stream(self, chunk_data, chunk_start):
+        with self.stream_lock:
+            self.stream.seek(self.stream_start + (chunk_start - self.start_index))
+            self.stream.write(chunk_data)
+
+
+class SequentialChunkDownloader(_ChunkDownloader):
+    def _update_progress(self, length):
+        self.progress_total += length
+
+    def _write_to_stream(self, chunk_data, chunk_start):
+        # chunk_start is ignored in the case of sequential download since we cannot seek the destination stream
+        self.stream.write(chunk_data)
+
 
 class StorageStreamDownloader(object):  # pylint: disable=too-many-instance-attributes
     """A streaming object to download from Azure Storage.
@@ -157,14 +219,17 @@ class StorageStreamDownloader(object):  # pylint: disable=too-many-instance-attr
     """
 
     def __init__(
-            self, service=None,
-            config=None,
-            offset=None,
-            length=None,
-            validate_content=None,
-            encryption_options=None,
-            **kwargs):
-        self.service = service
+        self,
+        client=None,
+        config=None,
+        offset=None,
+        length=None,
+        validate_content=None,
+        encryption_options=None,
+        extra_properties=None,
+        **kwargs
+    ):
+        self.client = client
         self.config = config
         self.offset = offset
         self.length = length
@@ -173,15 +238,13 @@ class StorageStreamDownloader(object):  # pylint: disable=too-many-instance-attr
         self.request_options = kwargs
         self.location_mode = None
         self._download_complete = False
-        self._current_content = None
-        self._iter_downloader = None
-        self._iter_chunks = None
 
         # The service only provides transactional MD5s for chunks under 4MB.
         # If validate_content is on, get only self.MAX_CHUNK_GET_SIZE for the first
         # chunk so a transactional MD5 can be retrieved.
-        self.first_get_size = self.config.max_single_get_size if not self.validate_content \
-            else self.config.max_chunk_get_size
+        self.first_get_size = (
+            self.config.max_single_get_size if not self.validate_content else self.config.max_chunk_get_size
+        )
         initial_request_start = self.offset if self.offset is not None else 0
         if self.length is not None and self.length - self.offset < self.first_get_size:
             initial_request_end = self.length
@@ -189,63 +252,12 @@ class StorageStreamDownloader(object):  # pylint: disable=too-many-instance-attr
             initial_request_end = initial_request_start + self.first_get_size - 1
 
         self.initial_range, self.initial_offset = process_range_and_offset(
-            initial_request_start, initial_request_end, self.length, self.encryption_options)
+            initial_request_start, initial_request_end, self.length, self.encryption_options
+        )
+
         self.download_size = None
         self.file_size = None
-        self.response = None
-        self.properties = None
-
-    def __len__(self):
-        return self.download_size
-
-    def __iter__(self):
-        raise TypeError("Async stream must be iterated asynchronously.")
-
-    def __aiter__(self):
-        return self
-
-    async def __anext__(self):
-        """Iterate through responses."""
-        if self._current_content is None:
-            if self.download_size == 0:
-                self._current_content = b""
-            else:
-                self._current_content = await process_content(
-                    self.response, self.initial_offset[0], self.initial_offset[1], self.encryption_options)
-            if not self._download_complete:
-                data_end = self.file_size
-                if self.length is not None:
-                    # Use the length unless it is over the end of the file
-                    data_end = min(self.file_size, self.length + 1)
-                self._iter_downloader = _AsyncChunkDownloader(
-                    service=self.service,
-                    total_size=self.download_size,
-                    chunk_size=self.config.max_chunk_get_size,
-                    current_progress=self.first_get_size,
-                    start_range=self.initial_range[1] + 1,  # start where the first download ended
-                    end_range=data_end,
-                    stream=None,
-                    parallel=False,
-                    validate_content=self.validate_content,
-                    encryption_options=self.encryption_options,
-                    use_location=self.location_mode,
-                    **self.request_options)
-                self._iter_chunks = self._iter_downloader.get_chunk_offsets()
-        elif self._download_complete:
-            raise StopAsyncIteration("Download complete")
-        else:
-            try:
-                chunk = next(self._iter_chunks)
-            except StopIteration:
-                raise StopAsyncIteration("DownloadComplete")
-            self._current_content = await self._iter_downloader.yield_chunk(chunk)
-
-        return self._current_content
-
-    async def setup(self, extra_properties=None):
-        if self.response:
-            raise ValueError("Download stream already initialized.")
-        self.response = await self._initial_request()
+        self.response = self._initial_request()
         self.properties = self.response.properties
 
         # Set the content length to the download size instead of the size of
@@ -253,7 +265,7 @@ class StorageStreamDownloader(object):  # pylint: disable=too-many-instance-attr
         self.properties.size = self.download_size
 
         # Overwrite the content range to the user requested range
-        self.properties.content_range = 'bytes {0}-{1}/{2}'.format(self.offset, self.length, self.file_size)
+        self.properties.content_range = "bytes {0}-{1}/{2}".format(self.offset, self.length, self.file_size)
 
         # Set additional properties according to download type
         if extra_properties:
@@ -265,22 +277,62 @@ class StorageStreamDownloader(object):  # pylint: disable=too-many-instance-attr
         # TODO: Set to the stored MD5 when the service returns this
         self.properties.content_md5 = None
 
-    async def _initial_request(self):
+    def __len__(self):
+        return self.download_size
+
+    def __iter__(self):
+        if self.download_size == 0:
+            content = b""
+        else:
+            content = process_content(
+                self.response, self.initial_offset[0], self.initial_offset[1], self.encryption_options
+            )
+
+        if content is not None:
+            yield content
+        if self._download_complete:
+            return
+
+        data_end = self.file_size
+        if self.length is not None:
+            # Use the length unless it is over the end of the file
+            data_end = min(self.file_size, self.length + 1)
+
+        downloader = SequentialChunkDownloader(
+            client=self.client,
+            total_size=self.download_size,
+            chunk_size=self.config.max_chunk_get_size,
+            current_progress=self.first_get_size,
+            start_range=self.initial_range[1] + 1,  # start where the first download ended
+            end_range=data_end,
+            stream=None,
+            validate_content=self.validate_content,
+            encryption_options=self.encryption_options,
+            use_location=self.location_mode,
+            **self.request_options
+        )
+
+        for chunk in downloader.get_chunk_offsets():
+            yield downloader.yield_chunk(chunk)
+
+    def _initial_request(self):
         range_header, range_validation = validate_and_format_range_headers(
             self.initial_range[0],
             self.initial_range[1],
             start_range_required=False,
             end_range_required=False,
-            check_content_md5=self.validate_content)
+            check_content_md5=self.validate_content
+        )
 
         try:
-            location_mode, response = await self.service.download(
+            location_mode, response = self.client.download(
                 range=range_header,
                 range_get_content_md5=range_validation,
                 validate_content=self.validate_content,
                 data_stream_total=None,
                 download_stream_current=0,
-                **self.request_options)
+                **self.request_options
+            )
 
             # Check the location we read from to ensure we use the same one
             # for subsequent requests.
@@ -303,11 +355,12 @@ class StorageStreamDownloader(object):  # pylint: disable=too-many-instance-attr
                 # request a range, do a regular get request in order to get
                 # any properties.
                 try:
-                    _, response = await self.service.download(
+                    _, response = self.client.download(
                         validate_content=self.validate_content,
                         data_stream_total=0,
                         download_stream_current=0,
-                        **self.request_options)
+                        **self.request_options
+                    )
                 except HttpResponseError as error:
                     process_storage_error(error)
 
@@ -321,14 +374,15 @@ class StorageStreamDownloader(object):  # pylint: disable=too-many-instance-attr
         # If file size is large, download the rest of the file in chunks.
         if response.properties.size != self.download_size:
             # Lock on the etag. This can be overriden by the user by specifying '*'
-            if self.request_options.get('modified_access_conditions'):
-                if not self.request_options['modified_access_conditions'].if_match:
-                    self.request_options['modified_access_conditions'].if_match = response.properties.etag
+            if self.request_options.get("modified_access_conditions"):
+                if not self.request_options["modified_access_conditions"].if_match:
+                    self.request_options["modified_access_conditions"].if_match = response.properties.etag
         else:
             self._download_complete = True
+
         return response
 
-    async def content_as_bytes(self, max_concurrency=1):
+    def content_as_bytes(self, max_concurrency=1):
         """Download the contents of this file.
 
         This operation is blocking until all data is downloaded.
@@ -338,10 +392,10 @@ class StorageStreamDownloader(object):  # pylint: disable=too-many-instance-attr
         :rtype: bytes
         """
         stream = BytesIO()
-        await self.download_to_stream(stream, max_concurrency=max_concurrency)
+        self.download_to_stream(stream, max_concurrency=max_concurrency)
         return stream.getvalue()
 
-    async def content_as_text(self, max_concurrency=1, encoding='UTF-8'):
+    def content_as_text(self, max_concurrency=1, encoding="UTF-8"):
         """Download the contents of this file, and decode as text.
 
         This operation is blocking until all data is downloaded.
@@ -350,10 +404,10 @@ class StorageStreamDownloader(object):  # pylint: disable=too-many-instance-attr
             The number of parallel connections with which to download.
         :rtype: str
         """
-        content = await self.content_as_bytes(max_concurrency=max_concurrency)
+        content = self.content_as_bytes(max_concurrency=max_concurrency)
         return content.decode(encoding)
 
-    async def download_to_stream(self, stream, max_concurrency=1):
+    def download_to_stream(self, stream, max_concurrency=1):
         """Download the contents of this file to a stream.
 
         :param stream:
@@ -363,12 +417,8 @@ class StorageStreamDownloader(object):  # pylint: disable=too-many-instance-attr
         :returns: The properties of the downloaded file.
         :rtype: Any
         """
-        if self._iter_downloader:
-            raise ValueError("Stream is currently being iterated.")
-
         # the stream must be seekable if parallel download is required
-        parallel = max_concurrency > 1
-        if parallel:
+        if max_concurrency > 1:
             error_message = "Target stream handle must be seekable."
             if sys.version_info >= (3,) and not stream.seekable():
                 raise ValueError(error_message)
@@ -381,8 +431,9 @@ class StorageStreamDownloader(object):  # pylint: disable=too-many-instance-attr
         if self.download_size == 0:
             content = b""
         else:
-            content = await process_content(
-                self.response, self.initial_offset[0], self.initial_offset[1], self.encryption_options)
+            content = process_content(
+                self.response, self.initial_offset[0], self.initial_offset[1], self.encryption_options
+            )
 
         # Write the content to the user stream
         if content is not None:
@@ -395,37 +446,30 @@ class StorageStreamDownloader(object):  # pylint: disable=too-many-instance-attr
             # Use the length unless it is over the end of the file
             data_end = min(self.file_size, self.length + 1)
 
-        downloader = _AsyncChunkDownloader(
-            service=self.service,
+        downloader_class = ParallelChunkDownloader if max_concurrency > 1 else SequentialChunkDownloader
+        downloader = downloader_class(
+            client=self.client,
             total_size=self.download_size,
             chunk_size=self.config.max_chunk_get_size,
             current_progress=self.first_get_size,
             start_range=self.initial_range[1] + 1,  # start where the first download ended
             end_range=data_end,
             stream=stream,
-            parallel=parallel,
             validate_content=self.validate_content,
             encryption_options=self.encryption_options,
             use_location=self.location_mode,
-            **self.request_options)
+            **self.request_options
+        )
 
-        dl_tasks = downloader.get_chunk_offsets()
-        running_futures = [
-            asyncio.ensure_future(downloader.process_chunk(d))
-            for d in islice(dl_tasks, 0, max_concurrency)
-        ]
-        while running_futures:
-            # Wait for some download to finish before adding a new one
-            _done, running_futures = await asyncio.wait(
-                running_futures, return_when=asyncio.FIRST_COMPLETED)
-            try:
-                next_chunk = next(dl_tasks)
-            except StopIteration:
-                break
-            else:
-                running_futures.add(asyncio.ensure_future(downloader.process_chunk(next_chunk)))
+        if max_concurrency > 1:
+            import concurrent.futures
+            executor = concurrent.futures.ThreadPoolExecutor(max_concurrency)
+            list(executor.map(
+                    with_current_context(downloader.process_chunk),
+                    downloader.get_chunk_offsets()
+                ))
+        else:
+            for chunk in downloader.get_chunk_offsets():
+                downloader.process_chunk(chunk)
 
-        if running_futures:
-            # Wait for the remaining downloads to finish
-            await asyncio.wait(running_futures)
         return self.properties
