+++ sdk/storage/azure-storage-file/azure/storage/file/file_client.py
@@ -26,7 +26,6 @@ from ._generated import AzureFileStorage
 from ._generated.version import VERSION
 from ._generated.models import StorageErrorException, FileHTTPHeaders
 from ._shared.uploads import IterStreamer, FileChunkUploader, upload_data_chunks
-from ._shared.downloads import StorageStreamDownloader
 from ._shared.base_client import StorageAccountHostsMixin, parse_connection_str, parse_query
 from ._shared.request_handlers import add_metadata_headers, get_length
 from ._shared.response_handlers import return_response_headers, process_storage_error
@@ -36,6 +35,7 @@ from ._deserialize import deserialize_file_properties, deserialize_file_stream
 from ._polling import CloseHandles
 from .models import HandlesPaged, NTFSAttributes  # pylint: disable=unused-import
 from ._shared_access_signature import FileSharedAccessSignature
+from .download import StorageStreamDownloader
 
 if TYPE_CHECKING:
     from datetime import datetime
@@ -649,7 +649,7 @@ class FileClient(StorageAccountHostsMixin):
             raise ValueError("Offset value must not be None if length is set.")
 
         return StorageStreamDownloader(
-            service=self._client.file,
+            client=self._client.file,
             config=self._config,
             offset=offset,
             length=length,
